indigo/backend/pyproject.toml
[tool.poetry]
name = "indigo-app"
version = "0.1.0"
description = "ADLM 2025 Data Challenge Contextual Retrieval Application"
authors = ["Jon Montgomery <jmontgom543@gmail.com>"]
packages = [{ include = "app" }]
license = "MIT"
repository = "https://github.com/jonmontg/ADLM-2025-Data-Challenge"

[tool.poetry.dependencies]
python = "3.11.13"
PyMuPDF = "^1.23.0"
pytesseract = "^0.3.10"
tqdm = "^4.66.0"
scikit-learn = "^1.3.0"
numpy = "2.3.2"
chonkie = {extras = ["st"], version = "^1.2.1"}
chonky = "^0.1.6"
boto3 = "^1.34.0"
pyspark = "^3.5.0"
pyarrow = "^21.0.0"
polars = "^1.33.1"
ftfy = "^6.3.1"
rank-bm25 = "^0.2.2"
nltk = "^3.9.1"
joblib = "^1.5.2"
zstandard = "^0.24.0"
orjson = "^3.11.3"
faiss-cpu = "^1.12.0"
openai = "^1.107.2"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"

[tool.poetry.scripts]
build = "app.build:main"
main = "app.main:main"
app = "app.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
